+++ packages/tswebextension/CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Recovered work of the blocking scriptlets `click2load.html`
+
+### Changed
+- Updated `@adguard/scriptlets` to `v1.7.20`
+
+
 ## [0.0.42]
+
 ### Changed
 - Fix js and css injection error handling
 - Fix request events initialization